% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ACECor}
\alias{ACECor}
\title{Correlation Matrix for ACE Algorithm Performance}
\usage{
ACECor(mirexace)
}
\arguments{
\item{mirexace}{a \code{mirexace} object}
}
\description{
\code{ACECor} returns the inter-correlation matrix of algorithm
performance for a MIREX ACE evaluation. Correlations appear in the
lower triangle and p-values appear in the upper triangle, corrected
for multiple comparisons.
}

