% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ACEOutliers}
\alias{ACEOutliers}
\title{Identify outliers in MIREX ACE performance according to Chauvenet's
criterion.}
\usage{
ACEOutliers(mirexace)
}
\description{
Identify outliers in MIREX ACE performance according to Chauvenet's
criterion.
}

