% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mirexace}
\alias{plot.mirexace}
\title{CLDs and Boxplots for ACE Performance}
\usage{
\method{plot}{mirexace}(x, level = 0.005, ...)
}
\arguments{
\item{x}{a \code{mirexace} object}

\item{level}{significance level for compact letter display}

\item{...}{graphical arguments passed on to \code{plot}}
}
\description{
Plots a compact letter display with weighted response boxplots for
MIREX ACE results.
}

