% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotACEClust}
\alias{PlotACEClust}
\title{Plots a hierarchical clustering of MIREX ACE algorithms based on their
correlations in performance.}
\usage{
PlotACEClust(mirexace, ...)
}
\description{
Plots a hierarchical clustering of MIREX ACE algorithms based on their
correlations in performance.
}

