% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadACEResults}
\alias{ReadACEResults}
\title{MIREX ACE results for a particular chord vocabulary or segmentation}
\usage{
ReadACEResults(results.directory, analysis)
}
\arguments{
\item{results.directory}{directory containing the results in Johan
Pauwels's output format}

\item{analysis}{chord vocabulary or segmentation to load}
}
\description{
MIREX ACE results for a particular chord vocabulary or segmentation
}

