% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ACECor}
\alias{ACECor}
\title{Correlation matrix of algorithm performance for MIREX ACE.}
\usage{
ACECor(mirexace)
}
\description{
Correlations appear in the lower triangle and p-values appear in the
upper triangle, corrected for multiple comparisons.
}

