% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ACEOutliers}
\alias{ACEOutliers}
\title{Outliers in ACE Evaluations}
\usage{
ACEOutliers(mirexace)
}
\arguments{
\item{mirexace}{a \code{mirexace} object}
}
\description{
\code{ACEOutliers} returns outliers in MIREX ACE performance
according to Chauvenet's criterion.
}

