% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotACEClust}
\alias{PlotACEClust}
\title{Hierarchical Clustering of ACE Algorithms}
\usage{
PlotACEClust(mirexace, ...)
}
\arguments{
\item{mirexace}{a \code{mirexace} object}

\item{...}{graphical arguments passed on to \code{plot}}
}
\description{
\code{PlotACEClust} plots a hierarchical clustering of MIREX ACE
algorithms based on their inter-correlations in performance.
}

