% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadACEResults}
\alias{ReadACEResults}
\title{Read MIREX ACE Results}
\usage{
ReadACEResults(results.directory, analysis)
}
\arguments{
\item{results.directory}{directory containing the results in Johan
Pauwels's output format}

\item{analysis}{chord vocabulary or segmentation to load}
}
\description{
\code{ReadACEResults} reads MIREX audio chord estimation results
for a particular chord vocabulary or segmentation from the output
of Johan Pauwels's evaluation scripts.
}

